pyproject.toml
[project]
name = "parlar"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai[realtime]>=1.102.0",
    "sounddevice>=0.5.2",
    "python-dotenv>=1.0.1",
    "rich>=13.7.1",
]
